+++ python/src/pyxspress/create_config/modules/create_odin_launch_script.py
@@ -1,3 +1,7 @@
+import os
+from pathlib import Path
+
+
 def _odin_processes(num_cards):
     odin_process_string = "procServ -P 4001 /odin/config/stOdinServer.sh\n"
     odin_process_string += "procServ -P 4002 /odin/config/stMetaWriter.sh\n"
@@ -18,22 +22,29 @@ def _odin_processes(num_cards):
     return odin_process_string
 
 
-def launch_n_chan(num_cards, num_chans, template_dir, files_dir):
+def launch_n_chan(
+    num_cards: int, num_chans: int, template_dir: Path, target_dir: Path
+) -> None:
+    """Create the launch shell script file
+
+    Args:
+        num_cards (int): Number of cards in Xspress system
+        num_chans (int): Number of channels in Xspress system
+        template_dir (Path): Template directory
+        target_dir (Path): Output directory
+    """
     odin_processes_str = _odin_processes(num_cards)
     with open(template_dir / "launch_nchan_odin.sh.template", "r") as launch_temp:
         launch_file_string = launch_temp.read()
-        launch_file_string = launch_file_string.replace("{chans}", str(num_chans))
-        launch_file_string = launch_file_string.replace(
-            "{odin_processes}", odin_processes_str
-        )
-    with open(files_dir / f"launch_{num_chans}chan_odin.sh", "w") as launch_file:
+
+    launch_file_string = launch_file_string.replace("{chans}", str(num_chans))
+    launch_file_string = launch_file_string.replace(
+        "{odin_processes}", odin_processes_str
+    )
+
+    target_filepath = target_dir / f"launch_{num_chans}chan_odin.sh"
+    with open(target_filepath, "w") as launch_file:
         launch_file.write(launch_file_string)
 
-
-def yaml(num_cards, template_dir, files_dir):
-    processes = (num_cards * 2) + 4
-    with open(template_dir / "odin_proc_serv_ioc.yaml.template", "r") as yaml_temp:
-        yaml_string = yaml_temp.read()
-        yaml_string = yaml_string.replace("{processes}", str(processes))
-    with open(files_dir / "odin_proc_serv_ioc.yaml", "w") as yaml_file:
-        yaml_file.write(yaml_string)
+    # Make executable
+    os.chmod(target_filepath, 0o755)